% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_extraction_module.R
\name{data_diagnosis_f}
\alias{data_diagnosis_f}
\title{Saldae : data diagnosis}
\usage{
data_diagnosis_f(tisefka = NULL, categoricals_ukud = NULL,
  ukud_unit = NULL)
}
\arguments{
\item{tisefka}{raw data}

\item{categoricals_ukud}{time_based gruping variables}

\item{ukud_unit}{time unit}
}
\value{
list containing detailed diagnos (quality , comleteness outliers,...)
}
\description{
draw a diagnosis on data  from quality and statitistics perspectives.
}
\author{
Farid Azouaou
}
