% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_extraction_module.R
\name{ghred_tisefka_aqerru}
\alias{ghred_tisefka_aqerru}
\title{Read data main function}
\usage{
ghred_tisefka_aqerru(input_file = NULL, tala = NULL, tawriqt = NULL)
}
\arguments{
\item{input_file}{input file}

\item{tala}{data source (CSV/Excel/Database)}

\item{tawriqt}{excel sheet name}
}
\value{
tbl_df object
}
\description{
Read data main function
}
\author{
Farid Azouaou
}
