% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_extraction_module.R
\name{ghred_tisefka_csv}
\alias{ghred_tisefka_csv}
\title{Read data from inputfile (CSV)}
\usage{
ghred_tisefka_csv(input_file = NULL)
}
\arguments{
\item{input_file}{input file}
}
\value{
tbl_df object
}
\description{
Read data from inputfile (CSV)
}
\author{
Farid Azouaou
}
