% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_extraction_module.R
\name{ghred_tisefka_excel}
\alias{ghred_tisefka_excel}
\title{Read data from inputfile (Excel)}
\usage{
ghred_tisefka_excel(input_file = NULL, tawriqt = NULL)
}
\arguments{
\item{input_file}{inputfile object}

\item{tawriqt}{excel sheet name}
}
\value{
tbl_df object
}
\description{
Read data from inputfile (Excel)
}
\author{
Farid Azouaou
}
