% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_helpers.R
\name{initialize_helper}
\alias{initialize_helper}
\title{initilaize helper module}
\usage{
initialize_helper(target_folder = "helpers")
}
\arguments{
\item{target_folder}{the folder to use as a helpers}
}
\description{
it activate all the necessary element to help the user to understand the app components
}
\author{
Farid Azouaou
}
