% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SA_ml_app_creation.R
\name{SA_ML_APP_server}
\alias{SA_ML_APP_server}
\title{Saldae Dashboard Module Server Analytics}
\usage{
SA_ML_APP_server(input, output, session, tisefka = NULL, create_flag = TRUE)
}
\arguments{
\item{input}{input shinydashboard elements containing information to use for output generation}

\item{output}{output shinydashboard element}

\item{session}{shiny session}

\item{tisefka}{reactive object containing data}

\item{div_width}{dimension information about the framework(html object)}
}
\value{
output objects to be displayed in corresponding UI module
}
\description{
Saldae Dashboard module SERVER : render and generate multiple output objects for analytics
}
\author{
Farid Azouaou
}
