% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SaldaeTidyModels_v1.R
\name{SA_ml_engine_lm1}
\alias{SA_ml_engine_lm1}
\title{prepare data : split
initializing the models
prepare formula}
\usage{
SA_ml_engine_lm1(
  tisefka = tisefka,
  target_variable = NULL,
  ml_algo = "linear regression",
  explaining_variable = NULL,
  train_prop = 0.75,
  sa_scale = FALSE,
  pred_mode = "regression"
)
}
\description{
prepare a recipe
embed the recipe to the engine
fit the model
predict
}
\details{
model performances evaluation
}
