% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SaldaeTidyModels_v1.R
\name{SA_ml_performance_f}
\alias{SA_ml_performance_f}
\title{Saldae Performance evaluation function}
\usage{
SA_ml_performance_f(SA_test_predicted = NULL,
  pred_mode = "classification", target_variable = NULL)
}
\arguments{
\item{SA_test_predicted}{data frame containg predicted values and benchmark values}

\item{pred_mode}{prediction mode (classification or regression)}
}
\value{
a daa frame containing performances results (AUC , RMSE , MSE , MAE , MAPE)
}
\description{
t.b.d
}
\author{
Farid Azouaou
}
