% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SaldaeTidyModels_v1.R
\name{SA_tisefka_bdu}
\alias{SA_tisefka_bdu}
\title{Saldae ML model builder}
\usage{
SA_tisefka_bdu(
  tisefka = NULL,
  train_prop = 0.8,
  strat_class = NULL,
  rand_sample = NULL
)
}
\arguments{
\item{train_prop}{training proportion}

\item{strat_class}{stratification class}

\item{tisekfa}{prepared data for ML training}
}
\value{
trained model
}
\description{
t.b.d
}
\author{
Farid Azouaou
}
