% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_extraction_module.R
\name{SA_tisefka_impute}
\alias{SA_tisefka_impute}
\title{Saldae Missing Values Imputation:}
\usage{
SA_tisefka_impute(tisefka = NULL)
}
\arguments{
\item{tisefka}{original data (Dataframe or tbl object)}
}
\value{
cleaned data (imputed missing values)
}
\description{
detect and impute missing values
}
\author{
Farid Azouaou
}
